backups/requirementsworkingversion.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
et_xmlfile==2.0.0
evernote-backup==1.9.3
evernote3==1.25.14
Flask==3.1.0
google-auth==2.38.0
google-cloud-tasks
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
gunicorn==20.1.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
google-api-python-client==2.70.0
google-auth-httplib2
MarkupSafe==3.0.2
numpy==2.2.3
oauth2==1.9.0.post1
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
google-api-python-client

pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.13.0
